+++ locals.tf
@@ -5,6 +5,7 @@ locals {
       "hosts" = merge(
         local.instances_aws_transform,
         local.instances_gcp_transform,
+        local.instances_azr_transform,
         local.instances_var_transform
       )
     }
@@ -24,5 +25,10 @@ locals {
   # transform gcp instances object into expected structure
   instances_gcp_transform = {
     for instance in var.instances_gcp : try(instance.tags[0], instance.instance_id) => merge({ "ansible_host" = instance.network_interface.0.network_ip }, { for tag in instance.tags : regexall("[-\\w]+", tag)[0] => regexall("[-\\w]+", tag)[1] if length(regexall("[-\\w]+", tag)) == 2 })
+  }
+
+  # transform azr instances object into expected structure
+  instances_azr_transform = {
+    for instance in var.instances_azr : lookup(instance.tags, "Name", instance.id) => merge({ "ansible_host" = instance.private_ip_address }, instance.tags)
   }
 }